Refactor xlsx gen
